+++ str_1.c
@@ -1,6 +1,6 @@
 #include "shell.h"
 
-int strlen_alt(const char *s);
+size_t strlen_alt(const char *s);
 char *strcat_alt(char *dest, char *src);
 int strcmp_alt(char *s1, char *s2);
 char *strchr_alt(char *s, char c);
@@ -12,9 +12,9 @@ char *strcpy_alt(char *destination, char *source);
 *
 * Return: the length of given string 's'
 */
-int strlen_alt(const char *s)
+size_t strlen_alt(const char *s)
 {
-	int len = 0;
+	size_t len = 0;
 
 	while (s[len] != '\0')
 	{
